requirements.txt
cookiecutter
pygithub
pygit2